==> notebooks/text_classification/3 - from df to torch geometric data.ipynb <==
!pip install torch-geometric

from google.colab import drive
drive.mount('/content/drive/')

node_path = '/content/drive/MyDrive/node_classification/text_classification'

from torch_geometric.data import Data
import numpy as np
import torch

import pandas as pd
df_total = pd.read_pickle(f"{node_path}/data/total_data.pkl")
dict_edges = pd.read_pickle(f"{node_path}/data/edges_data.pkl")

dict_edges['edges_list'] = dict_edges['edges'].apply(lambda x: list(x))

df_total['split'].unique()

df_train = df_total[df_total['split'] == "train"].reset_index(drop=True)
df_test = df_total[df_total['split'] == "test"].reset_index(drop=True)
df_val = df_total[df_total['split'] == "val"].reset_index(drop=True)

len(df_train), len(df_test), len(df_val)

df_train.head(3)

# Setting up the data for pytorch geometric

---

x = torch.tensor(np.stack(df_total['embedding_normalized'].values))
edge_index = torch.tensor(np.stack(dict_edges['edges_list'].values),  dtype=torch.long).t().contiguous()
y = torch.tensor(df_total["label"].values)

data = Data(x=x, edge_index=edge_index, y=y)

data.num_classes = len(torch.unique(data.y))

train_indices = list(df_train['id'].values)
val_indices = list(df_val['id'].values)
test_indices = list(df_test['id'].values)

train_mask = torch.zeros(len(df_total), dtype=torch.bool)
val_mask = torch.zeros(len(df_total), dtype=torch.bool)
test_mask = torch.zeros(len(df_total), dtype=torch.bool)

train_mask[train_indices] = True
val_mask[val_indices] = True
test_mask[test_indices] = True

# Add masks to the Data object
data.train_mask = train_mask
data.val_mask = val_mask
data.test_mask = test_mask

print(data)

torch.save(data, f"{node_path}/data/ptg_complete_data.pth")
